ICMSLST-410 Merge ImportCase and ImportApplication
ImportCase and ImportApplication are technically a single entity, having
them as separate models leads to a struggle in development. ImportCase
and ImportApplicatoin are now merged as a single ImportApplication
model.
