+++ web/domains/case/_import/models.py
@@ -2,10 +2,92 @@ from django.db import models
 
 from web.domains.case.fir.models import FurtherInformationRequest
 from web.domains.case.models import CaseNote, UpdateRequest, VariationRequest
+from web.domains.commodity.models import CommodityGroup, CommodityType
+from web.domains.country.models import Country, CountryGroup
+from web.domains.importer.models import Importer
+from web.domains.office.models import Office
+from web.domains.template.models import Template
 from web.domains.user.models import User
 
 
-class ImportCase(models.Model):
+class ImportApplicationType(models.Model):
+    is_active = models.BooleanField(blank=False, null=False)
+    type_code = models.CharField(max_length=30, blank=False, null=False)
+    type = models.CharField(max_length=70, blank=False, null=False)
+    sub_type_code = models.CharField(max_length=30, blank=False, null=False)
+    sub_type = models.CharField(max_length=70, blank=True, null=True)
+    licence_type_code = models.CharField(max_length=20, blank=False, null=False)
+    sigl_flag = models.BooleanField(blank=False, null=False)
+    chief_flag = models.BooleanField(blank=False, null=False)
+    chief_licence_prefix = models.CharField(max_length=10, blank=True, null=True)
+    paper_licence_flag = models.BooleanField(blank=False, null=False)
+    electronic_licence_flag = models.BooleanField(blank=False, null=False)
+    cover_letter_flag = models.BooleanField(blank=False, null=False)
+    cover_letter_schedule_flag = models.BooleanField(blank=False, null=False)
+    category_flag = models.BooleanField(blank=False, null=False)
+    sigl_category_prefix = models.CharField(max_length=100, blank=True, null=True)
+    chief_category_prefix = models.CharField(max_length=10, blank=True, null=True)
+    default_licence_length_months = models.IntegerField(blank=True, null=True)
+    endorsements_flag = models.BooleanField(blank=False, null=False)
+    default_commodity_desc = models.CharField(max_length=200, blank=True, null=True)
+    quantity_unlimited_flag = models.BooleanField(blank=False, null=False)
+    unit_list_csv = models.CharField(max_length=200, blank=True, null=True)
+    exp_cert_upload_flag = models.BooleanField(blank=False, null=False)
+    supporting_docs_upload_flag = models.BooleanField(blank=False, null=False)
+    multiple_commodities_flag = models.BooleanField(blank=False, null=False)
+    guidance_file_url = models.CharField(max_length=4000, blank=True, null=True)
+    licence_category_description = models.CharField(max_length=1000, blank=True, null=True)
+
+    usage_auto_category_desc_flag = models.BooleanField(blank=False, null=False)
+    case_checklist_flag = models.BooleanField(blank=False, null=False)
+    importer_printable = models.BooleanField(blank=False, null=False)
+    origin_country_group = models.ForeignKey(
+        CountryGroup,
+        on_delete=models.PROTECT,
+        blank=True,
+        null=True,
+        related_name="import_application_types_from",
+    )
+    consignment_country_group = models.ForeignKey(
+        CountryGroup,
+        on_delete=models.PROTECT,
+        blank=True,
+        null=True,
+        related_name="import_application_types_to",
+    )
+    master_country_group = models.ForeignKey(
+        CountryGroup,
+        on_delete=models.PROTECT,
+        blank=True,
+        null=True,
+        related_name="import_application_types",
+    )
+    commodity_type = models.ForeignKey(
+        CommodityType, on_delete=models.PROTECT, blank=True, null=True
+    )
+    declaration_template = models.ForeignKey(
+        Template,
+        on_delete=models.PROTECT,
+        blank=False,
+        null=False,
+        related_name="declaration_application_types",
+    )
+    endorsements = models.ManyToManyField(Template, related_name="endorsement_application_types")
+    default_commodity_group = models.ForeignKey(
+        CommodityGroup, on_delete=models.PROTECT, blank=True, null=True
+    )
+
+    def __str__(self):
+        if self.sub_type:
+            return f"{self.type} ({self.sub_type})"
+        else:
+            return f"{self.type}"
+
+    class Meta:
+        ordering = ("type", "sub_type")
+
+
+class ImportApplication(models.Model):
     IN_PROGRESS = "IN_PROGRESS"
     SUBMITTED = "SUBMITTED"
     PROCESSING = "PROCESSING"
@@ -45,8 +127,12 @@ class ImportCase(models.Model):
     APPROVE = "APPROVE"
     DECISIONS = ((REFUSE, "Refuse"), (APPROVE, "Approve"))
 
+    is_active = models.BooleanField(blank=False, null=False, default=True)
     status = models.CharField(max_length=30, choices=STATUSES, blank=False, null=False)
     reference = models.CharField(max_length=50, blank=True, null=True)
+    applicant_reference = models.CharField(max_length=500, blank=True, null=True)
+    submit_datetime = models.DateTimeField(blank=True, null=True)
+    create_datetime = models.DateTimeField(blank=False, null=False, auto_now_add=True)
     variation_no = models.IntegerField(blank=False, null=False, default=0)
     legacy_case_flag = models.BooleanField(blank=False, null=False, default=False)
     chief_usage_status = models.CharField(
@@ -62,8 +148,74 @@ class ImportCase(models.Model):
     licence_end_date = models.DateField(blank=True, null=True)
     licence_extended_flag = models.BooleanField(blank=False, null=False, default=False)
     last_update_datetime = models.DateTimeField(blank=False, null=False, auto_now=True)
+    application_type = models.ForeignKey(
+        ImportApplicationType, on_delete=models.PROTECT, blank=False, null=False
+    )
+    submitted_by = models.ForeignKey(
+        User,
+        on_delete=models.PROTECT,
+        blank=True,
+        null=True,
+        related_name="submitted_import_applications",
+    )
+    created_by = models.ForeignKey(
+        User,
+        on_delete=models.PROTECT,
+        blank=False,
+        null=False,
+        related_name="created_import_applications",
+    )
     last_updated_by = models.ForeignKey(
         User, on_delete=models.PROTECT, blank=False, null=False, related_name="updated_import_cases"
+    )
+    importer = models.ForeignKey(
+        Importer,
+        on_delete=models.PROTECT,
+        blank=False,
+        null=False,
+        related_name="import_applications",
+    )
+    agent = models.ForeignKey(
+        Importer,
+        on_delete=models.PROTECT,
+        blank=True,
+        null=True,
+        related_name="agent_import_applications",
+    )
+    importer_office = models.ForeignKey(
+        Office,
+        on_delete=models.PROTECT,
+        blank=False,
+        null=True,
+        related_name="office_import_applications",
+    )
+    agent_office = models.ForeignKey(
+        Office,
+        on_delete=models.PROTECT,
+        blank=True,
+        null=True,
+        related_name="agent_office_import_applications",
+    )
+    contact = models.ForeignKey(
+        User,
+        on_delete=models.PROTECT,
+        blank=True,
+        null=True,
+        related_name="contact_import_applications",
+    )
+    origin_country = models.ForeignKey(
+        Country,
+        on_delete=models.PROTECT,
+        blank=True,
+        null=True,
+        related_name="import_applications_from",
+    )
+    consignment_country = models.ForeignKey(
+        Country,
+        on_delete=models.PROTECT,
+        blank=True,
+        null=True,
+        related_name="import_applications_to",
     )
     variation_requests = models.ManyToManyField(VariationRequest)
     case_notes = models.ManyToManyField(CaseNote)
@@ -80,7 +232,9 @@ class ConstabularyEmail(models.Model):
     STATUSES = ((OPEN, "Open"), (CLOSED, "Closed"), (DRAFT, "Draft"))
 
     is_active = models.BooleanField(blank=False, null=False, default=True)
-    case = models.ForeignKey(ImportCase, on_delete=models.PROTECT, blank=False, null=False)
+    application = models.ForeignKey(
+        ImportApplication, on_delete=models.PROTECT, blank=False, null=False
+    )
     status = models.CharField(max_length=30, blank=False, null=False, default=DRAFT)
     email_cc_address_list = models.CharField(max_length=4000, blank=True, null=True)
     email_subject = models.CharField(max_length=100, blank=False, null=True)
