+++ web/models/__init__.py
@@ -1,4 +1,3 @@
-from web.domains.application._import.models import *  # NOQA
 from web.domains.case._import.models import *  # NOQA
 from web.domains.case.access.models import *  # NOQA
 from web.domains.case.export.models import *  # NOQA
