+++ web/tests/domains/case/_import/factory.py
@@ -1,7 +1,7 @@
 import factory
 import factory.fuzzy
 import random
-from web.domains.application._import.models import ImportApplicationType
+from web.domains.case._import.models import ImportApplicationType
 from web.tests.domains.template.factory import TemplateFactory
