+++ web/tests/domains/case/_import/test_forms.py
@@ -1,7 +1,7 @@
 import logging
 from django.test import RequestFactory, TestCase
 from django.urls import reverse_lazy
-from web.domains.application._import.forms import NewImportApplicationForm
+from web.domains.case._import.forms import NewImportApplicationForm
 from web.tests.domains.importer.factory import ImporterFactory
 from web.tests.domains.user.factory import UserFactory
 from web.tests.domains.office.factory import OfficeFactory
