+++ web/urls.py
@@ -22,6 +22,6 @@ urlpatterns = [
     path("exporter/", include("web.domains.exporter.urls")),
     path("access/", include("web.domains.case.access.urls", namespace="access")),
     path("fir/", include("web.domains.case.fir.urls", namespace="fir")),
-    path("import/", include("web.domains.application._import.urls")),
+    path("import/", include("web.domains.case._import.urls")),
     path("mailshot/", include("web.domains.mailshot.urls")),
 ]
